convert .ogg to .wav/.mp3
attempting to fix the following issue:
For some reason, the skin makes my computer lag, while other skins don't. What do you think is the cause of that, and can I fix it?

EDIT: It's not heavy lag. Just short intermittent lag spikes in between maps. It gets really annoying when the game chokes for you and you end up killing your chances of a good score because of it.

The simple solution is to read another skin, but I can't.